feat: add performance snapshot service and related calculations
- Introduced `buildPerformanceSnapshot` and `processChartData` functions to generate performance metrics from trades and daily logs.
- Added filtering capabilities for strategies and date ranges in performance snapshots.
- Implemented various calculations including equity curves, drawdown data, and rolling metrics.
- Enhanced unit tests for data loader and trading frequency card components.
- Updated package dependencies for testing libraries.
