+++ package.json
@@ -63,6 +63,8 @@
     "@eslint/eslintrc": "^3",
     "@jest/globals": "^30.2.0",
     "@tailwindcss/postcss": "^4",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
     "@types/jest": "^30.0.0",
     "@types/node": "^20",
     "@types/react": "^19",
@@ -73,6 +75,7 @@
     "fake-indexeddb": "^6.2.2",
     "husky": "^9.1.7",
     "jest": "^30.2.0",
+    "jest-environment-jsdom": "^29.7.0",
     "tailwindcss": "^4",
     "ts-jest": "^29.4.4",
     "tw-animate-css": "^1.4.0",
