+++ tests/integration/indexeddb-data-loader.test.ts
@@ -1,3 +1,5 @@
+/** @jest-environment node */
+
 /**
  * IndexedDB Integration Tests with Data Loader
  *
