+++ tests/unit/trading-frequency-card.test.tsx
@@ -2,36 +2,43 @@ import { render, screen } from "@testing-library/react";
 import { TradingFrequencyCard } from "@/components/risk-simulator/trading-frequency-card";
 import { Trade } from "@/lib/models/trade";
 
-// Mock trade generator
+const SUMMARY_TEST_ID = "frequency-summary";
+const RATE_TEST_ID = "frequency-rate";
+
 function createMockTrades(count: number, startDate: Date, endDate: Date): Trade[] {
+  if (count === 0) {
+    return [];
+  }
+
   const trades: Trade[] = [];
-  const timeSpan = endDate.getTime() - startDate.getTime();
+  const span = endDate.getTime() - startDate.getTime();
 
   for (let i = 0; i < count; i++) {
-    const tradeDate = new Date(startDate.getTime() + (timeSpan * i / count));
+    const progress = count === 1 ? 0 : i / (count - 1);
+    const date = new Date(startDate.getTime() + span * progress);
     trades.push({
       id: `trade-${i}`,
-      dateOpened: tradeDate,
-      dateClosed: tradeDate,
+      dateOpened: date,
+      dateClosed: date,
       timeOpened: "09:30:00",
       timeClosed: "10:30:00",
       ticker: "SPY",
       strategy: "Test Strategy",
-      quantity: 100,
+      quantity: 1,
       entryPrice: 100,
       exitPrice: 101,
-      pl: 100,
+      pl: 1,
       plPercent: 1,
-      commission: 1,
-      totalCommissions: 2,
-      openingCommissionsFees: 1,
-      closingCommissionsFees: 1,
+      commission: 0,
+      totalCommissions: 0,
+      openingCommissionsFees: 0,
+      closingCommissionsFees: 0,
       tradeType: "LONG",
-      marginUsed: 10000,
-      accountBalance: 100000,
-      accountPeak: 100000,
-      capitalEfficiency: 1,
-      kelly: 0.05,
+      marginUsed: 0,
+      accountBalance: 0,
+      accountPeak: 0,
+      capitalEfficiency: 0,
+      kelly: 0,
     });
   }
 
@@ -39,120 +46,56 @@ function createMockTrades(count: number, startDate: Date, endDate: Date): Trade[
 }
 
 describe("TradingFrequencyCard", () => {
-  it("should display correct frequency for regular trader (252 trades/year)", () => {
-    const startDate = new Date("2023-01-01");
-    const endDate = new Date("2024-01-01");
-    const trades = createMockTrades(252, startDate, endDate);
+  it("shows annual frequency details for a regular trader", () => {
+    const trades = createMockTrades(252, new Date("2023-01-01"), new Date("2024-01-01"));
 
     render(<TradingFrequencyCard trades={trades} tradesPerYear={252} />);
 
-    expect(screen.getByText("252")).toBeInTheDocument();
     expect(screen.getByText("trades/year")).toBeInTheDocument();
-    expect(screen.getByText(/21 trades\/month/)).toBeInTheDocument();
-    expect(screen.getByText("252 trades")).toBeInTheDocument();
-    expect(screen.getByText(/1\.0 year/)).toBeInTheDocument();
+    expect(screen.getByTestId(RATE_TEST_ID)).toHaveTextContent("21 trades/month");
+    expect(screen.getByTestId(SUMMARY_TEST_ID)).toHaveTextContent(/252 trades over/i);
   });
 
-  it("should display correct frequency for high-frequency trader", () => {
-    const startDate = new Date("2024-01-01");
-    const endDate = new Date("2024-01-31"); // 1 month
-    const trades = createMockTrades(800, startDate, endDate); // ~40 trades/day
+  it("shows daily rate for a high-frequency trader", () => {
+    const trades = createMockTrades(800, new Date("2024-01-01"), new Date("2024-01-31"));
 
     render(<TradingFrequencyCard trades={trades} tradesPerYear={10400} />);
 
-    expect(screen.getByText("10,400")).toBeInTheDocument();
-    expect(screen.getByText("trades/year")).toBeInTheDocument();
-    expect(screen.getByText(/40 trades\/day/)).toBeInTheDocument(); // 10400/260 trading days
-    expect(screen.getByText("800 trades")).toBeInTheDocument();
-    expect(screen.getByText(/1 month/)).toBeInTheDocument();
+    expect(screen.getByTestId(RATE_TEST_ID)).toHaveTextContent(/40 trades\/day/i);
+    expect(screen.getByTestId(SUMMARY_TEST_ID)).toHaveTextContent(/800 trades over/i);
   });
 
-  it("should display correct frequency for active trader", () => {
-    const startDate = new Date("2023-01-01");
-    const endDate = new Date("2023-06-30"); // 6 months
-    const trades = createMockTrades(1000, startDate, endDate);
+  it("shows weekly rate for an active trader", () => {
+    const trades = createMockTrades(1000, new Date("2023-01-01"), new Date("2023-06-30"));
 
     render(<TradingFrequencyCard trades={trades} tradesPerYear={2000} />);
 
-    expect(screen.getByText("2,000")).toBeInTheDocument();
-    expect(screen.getByText("trades/year")).toBeInTheDocument();
-    expect(screen.getByText(/38 trades\/week/)).toBeInTheDocument(); // 2000/52 weeks
-    expect(screen.getByText("1,000 trades")).toBeInTheDocument();
-    expect(screen.getByText(/6 months/)).toBeInTheDocument();
+    expect(screen.getByTestId(RATE_TEST_ID)).toHaveTextContent(/38 trades\/week/i);
+    expect(screen.getByTestId(SUMMARY_TEST_ID)).toHaveTextContent(/1,000 trades over/i);
   });
 
-  it("should display correct frequency for occasional trader", () => {
-    const startDate = new Date("2022-01-01");
-    const endDate = new Date("2024-01-01"); // 2 years
-    const trades = createMockTrades(100, startDate, endDate);
+  it("shows yearly rate for an occasional trader", () => {
+    const trades = createMockTrades(100, new Date("2022-01-01"), new Date("2024-01-01"));
 
     render(<TradingFrequencyCard trades={trades} tradesPerYear={50} />);
 
-    expect(screen.getByText("50")).toBeInTheDocument();
-    expect(screen.getByText("trades/year")).toBeInTheDocument();
-    expect(screen.getByText(/50 trades\/year/)).toBeInTheDocument();
-    expect(screen.getByText("100 trades")).toBeInTheDocument();
-    expect(screen.getByText(/2\.0 years/)).toBeInTheDocument();
+    expect(screen.getByTestId(RATE_TEST_ID)).toHaveTextContent(/50 trades\/year/i);
+    expect(screen.getByTestId(SUMMARY_TEST_ID)).toHaveTextContent(/100 trades over/i);
   });
 
-  it("should handle edge case with less than 2 trades", () => {
+  it("handles minimal trade history", () => {
     const trades = createMockTrades(1, new Date("2024-01-01"), new Date("2024-01-01"));
 
     render(<TradingFrequencyCard trades={trades} tradesPerYear={252} />);
 
-    expect(screen.getByText("252")).toBeInTheDocument();
-    expect(screen.getByText("1 trade")).toBeInTheDocument(); // Singular
+    expect(screen.getByTestId(SUMMARY_TEST_ID)).toHaveTextContent(/1 trade over/i);
+    expect(screen.getByTestId(RATE_TEST_ID)).toHaveTextContent(/21 trades\/month/i);
   });
 
-  it("should handle empty trades array", () => {
+  it("handles empty trade history", () => {
     render(<TradingFrequencyCard trades={[]} tradesPerYear={252} />);
 
-    expect(screen.getByText("252")).toBeInTheDocument();
-    expect(screen.getByText("0 trades")).toBeInTheDocument();
-  });
-
-  it("should display trading rate correctly for different frequencies", () => {
-    // Test formatTradingRate logic
-    const startDate = new Date("2024-01-01");
-    const endDate = new Date("2024-01-02");
-
-    // Very high frequency: 40+ trades per day
-    const trades = createMockTrades(40, startDate, endDate);
-    const { rerender } = render(<TradingFrequencyCard trades={trades} tradesPerYear={14600} />);
-    expect(screen.getByText(/40 trades\/day/)).toBeInTheDocument();
-
-    // Active weekly trader
-    rerender(<TradingFrequencyCard trades={trades} tradesPerYear={1040} />);
-    expect(screen.getByText(/20 trades\/week/)).toBeInTheDocument();
-
-    // Monthly trader
-    rerender(<TradingFrequencyCard trades={trades} tradesPerYear={240} />);
-    expect(screen.getByText(/20 trades\/month/)).toBeInTheDocument();
-
-    // Occasional trader
-    rerender(<TradingFrequencyCard trades={trades} tradesPerYear={50} />);
-    expect(screen.getByText(/50 trades\/year/)).toBeInTheDocument();
-  });
-
-  it("should format time period correctly", () => {
-    const startDate = new Date("2024-01-01");
-
-    // Test years display
-    let endDate = new Date("2025-06-01"); // 1.5 years
-    let trades = createMockTrades(100, startDate, endDate);
-    const { rerender } = render(<TradingFrequencyCard trades={trades} tradesPerYear={252} />);
-    expect(screen.getByText(/1\.4 years/)).toBeInTheDocument();
-
-    // Test months display
-    endDate = new Date("2024-04-01"); // 3 months
-    trades = createMockTrades(100, startDate, endDate);
-    rerender(<TradingFrequencyCard trades={trades} tradesPerYear={252} />);
-    expect(screen.getByText(/3 months/)).toBeInTheDocument();
-
-    // Test days display
-    endDate = new Date("2024-01-15"); // 14 days
-    trades = createMockTrades(100, startDate, endDate);
-    rerender(<TradingFrequencyCard trades={trades} tradesPerYear={252} />);
-    expect(screen.getByText(/14 days/)).toBeInTheDocument();
+    expect(screen.getByTestId(SUMMARY_TEST_ID)).toHaveTextContent(/0 trades over/i);
+    expect(screen.getByTestId(RATE_TEST_ID)).toHaveTextContent(/21 trades\/month/i);
   });
 });
